idx proto

+++ includes/corewar.h
@@ -169,7 +169,7 @@ size_t					revgetop(bool (*op)());
 void		ft_warcycle(t_cor *core, bool b_alive);
 void		ft_increase_cycle(t_cor *core);
 void		ft_browseprocess(t_cor *core);
-size_t		idx(t_process *proc, size_t jump);
+long int		idx(t_process *proc, long int jump);
 void		ft_determinejmpdist(t_cor *core, t_process *proc);
 size_t		ind(t_cor *core, t_process *proc, size_t PC);
 size_t		getparam(t_cor *core, t_process *proc, size_t param, size_t label);
